app/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0,<2.10
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
dotenv>=0.9.9
python-multipart>=0.0.6  # Required for FastAPI Form data handling

# Development and testing
pytest>=8.4.1
pytest-asyncio>=1.1.0
httpx>=0.27.0

# ASW Lambda
mangum==0.19.0

# Crypto
cryptography


# database module dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
aiosqlite>=0.20.0
greenlet
psycopg2-binary>=2.9.0  # For PostgreSQL support

# auth module dependencies
python-jose[cryptography]>=3.3.0

# aihub module dependencies
openai==2.16.0
sse-starlette>=1.6.0

# payment module dependencies
stripe>=12.0.0
